move watermark to common.ir.vertex.transform
